# Use multiple TypeScript projects