--- a/Makefile
+++ b/Makefile
@@ -60,13 +60,10 @@ code-quality: lint
 tscheck:
 	$(MAKEJS) tscheck
 
-tscheck-slow:
-	$(MAKEJS) tscheck-slow
-
 clean-ts:
 	$(MAKEJS) clean-ts
 
-lint-ci: lint check-compat-data
+lint-ci: lint-ci check-compat-data
 
 generate-readme:
 	$(NODE) scripts/generators/readmes.js
@@ -93,7 +90,7 @@ test: lint test-only
 clone-license:
 	$(MAKEJS) clone-license
 
-prepublish-prepare-dts:
+prepublish-prepare-dts: tscheck
 	$(MAKEJS) prepublish-prepare-dts
 
 prepublish-build: