--- a/Makefile.source.mjs
+++ b/Makefile.source.mjs
@@ -349,6 +349,9 @@ target["clone-license"] = function () {
  */
 
 target["lint"] = function () {
+  // Pre-run tscheck, otherwise typescript-eslint
+  // will run it by itsels
+  target["tscheck"]();
   env(
     () => {
       yarn(EslintArgs);