--- a/codemods/babel-plugin-codemod-object-assign-to-object-spread/tsconfig.json
+++ b/codemods/babel-plugin-codemod-object-assign-to-object-spread/tsconfig.json
@@ -12,9 +12,6 @@
   "references": [
     {
       "path": "../../packages/babel-helper-plugin-utils"
-    },
-    {
-      "path": "../../packages/babel-core"
     }
   ]
 }