--- a/codemods/babel-plugin-codemod-optional-catch-binding/tsconfig.json
+++ b/codemods/babel-plugin-codemod-optional-catch-binding/tsconfig.json
@@ -12,9 +12,6 @@
   "references": [
     {
       "path": "../../packages/babel-helper-plugin-utils"
-    },
-    {
-      "path": "../../packages/babel-core"
     }
   ]
 }