… Makefile.js
… eslint/babel-eslint-parser/tsconfig.json
… package.json
… packages/babel-cli/tsconfig.json
… packages/babel-code-frame/tsconfig.json
… packages/babel-compat-data/tsconfig.json
… packages/babel-core/package.json
… packages/babel-core/tsconfig.json
… packages/babel-helper-annotate-as-pure/tsconfig.json
… packages/babel-helper-builder-binary-assignment-operator-visitor/tsconfig.json
… packages/babel-helper-builder-react-jsx/tsconfig.json
… packages/babel-helper-create-class-features-plugin/tsconfig.json
… packages/babel-helper-create-regexp-features-plugin/tsconfig.json
… packages/babel-helper-fixtures/tsconfig.json
… packages/babel-helper-import-to-platform-api/tsconfig.json
… packages/babel-helper-member-expression-to-functions/tsconfig.json
… packages/babel-helper-optimise-call-expression/tsconfig.json
… packages/babel-helper-plugin-utils/package.json
… packages/babel-helper-plugin-utils/tsconfig.json
… packages/babel-helper-remap-async-to-generator/tsconfig.json
… packages/babel-helper-replace-supers/tsconfig.json
… packages/babel-helper-skip-transparent-expression-wrappers/tsconfig.json
… packages/babel-helper-string-parser/tsconfig.json
… packages/babel-helper-validator-identifier/tsconfig.json
… packages/babel-helper-validator-option/tsconfig.json
… packages/babel-helper-wrap-function/tsconfig.json
… packages/babel-highlight/tsconfig.json
… packages/babel-node/tsconfig.json
… packages/babel-plugin-bugfix-firefox-class-in-computed-class-key/tsconfig.json
… packages/babel-plugin-bugfix-safari-id-destructuring-collision-in-function-expression/tsconfig.json