<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | `Fixes #1, Fixes #2` <!-- remove the (`) quotes and write "Fixes" before the number to link the issues -->
| Patch: Bug Fix?          |
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->

With [Project References](https://www.typescriptlang.org/docs/handbook/project-references.html), `tsc` is able to only type-check "projects" whose contents changed, without having to re-check the whole monorepo every time.

Usually, when using a monorepo, there is a 1:1 mapping between packages and TypeScript projects. In our case this doesn't work, because Project References do not support cycles. Instead, we generate one project per strongly connected component in our packages graph: we currently have only one SCC[^1], and its `tsconfig.json` is in `packages/babel-helper-plugin-utils/tsconfig.json`.

All the `tsconfig.json` files are not manually maintained (except for `/tsconfig.base.json`), but they are generated via `node scripts/generators/tsconfig.js`.

See https://github.com/babel/babel/pull/16430#issuecomment-2070365786 for how this impacts our development workflows.

<details>
<summary>Old description</summary>

I also experimented with running TypeScript on multiple cores rather than in just one. `tsc` doesn't support it by default, but you can use `node scripts/parallel-tsc/tsc.js .` (or passing the path of any other project) to do it. It seems to be ~40% faster than plain `tsc` on a MacBook M2 with 8 cores (but for some reason only 6 are being used).

This PR makes running a full type checking significantly slower (this will affect CI), but it makes incremental builds significantly faster (up to 10x, depending on which package is modified, and this will affect local development). This are the results I'm getting now (but I was getting [different results](https://github.com/microsoft/TypeScript/issues/30511#issuecomment-2054037835) this morning, unless I hallucinate them):
- without this PR, `rm -rf dts tsconfig.tsbuildinfo && yarn tsc -b .` takes 31s (both with and without `skipLibCheck`)
- with this PR, `make clean-ts && yarn tsc -b .` takes 5m 22s
- with this PR and `skipLibCheck: true`, `make clean-ts && yarn tsc -b .` takes 1min 53s
- with this PR, `make clean-ts && node scripts/parallel-tsc/tsc.js .` takes 2m 42s
- with this PR and `skipLibCheck: true`, `make clean-ts && node scripts/parallel-tsc/tsc.js .` takes 1min 16s

I believe the reason `skipLibCheck: true` significantly improves performance is that it lets TS avoid re-validating all the `.d.ts` files it generated for the intermediate projects. I will mark this PR as ready once we can get the `skipLibCheck: true` results while still validating our manually-written `.d.td` files.

</details>

---

@jakebailey Here is the branch :) Please feel free to ask any question that might help you, if you are going to take a look

[^1]:  (`@babel/helper-plugin-utils <> @babel/code-frame <> @babel/compat-data <> @babel/core <> @babel/generator <> @babel/helper-check-duplicate-nodes <> @babel/helper-compilation-targets <> @babel/helper-environment-visitor <> @babel/helper-fixtures <> @babel/helper-function-name <> @babel/helper-hoist-variables <> @babel/helper-module-imports <> @babel/helper-module-transforms <> @babel/helper-plugin-test-runner <> @babel/helper-simple-access <> @babel/helper-split-export-declaration <> @babel/helper-string-parser <> @babel/helper-transform-fixture-test-runner <> @babel/helper-validator-identifier <> @babel/helper-validator-option <> @babel/helpers <> @babel/highlight <> @babel/parser <> @babel/template <> @babel/traverse <> @babel/types`, with 26 packages out of ~150, but I'm looking into reducing its size by splitting some packages across multiple projects to reduce cycles).