@nicolo-ribaudo 2024-04-18
The last commit (d826ec0799, `Make types faster`) significantly improves the situation by:
- generating some giant object types at build time rather than at type checking time
- replacing some `type` aliases with `interface` (that now we can do because the object shapes are statically known)

| | **Before** | **After** |
|-|-|-|
| `make clean-ts && yarn tsc -b .`, `skipLibCheck: false` | 5m 10s | 3m 38s |
| `make clean-ts && yarn tsc -b .`, `skipLibCheck: true`  | 1m 53s | 24s |
| `make clean-ts && node scripts/parallel-tsc/tsc.js .`, `skipLibCheck: false` | 1m 9s | 59s |
| `make clean-ts && node scripts/parallel-tsc/tsc.js .`, `skipLibCheck: true`  | 41s | 22s |