@nicolo-ribaudo 2024-04-19
By forcing `NodePath` to be distributive over its type argument, the type-checking without cache with `yarn tsc -b .` and `skipLibCheck: false` goes down from 3m38s to just 11s :)

By doing something like this: https://www.typescriptlang.org/play?#code/KYDwDg9gTgLgBAbwL4FgBQ6CWA7GwoBmAhgMbBwByEAJsAIKLpzNwwCeYwAXHAER28A3ExYkAFpgA21HlVoAhOAB9KNYAGFhaVBjQ48hUuTnBFCEc3aceveUItwSEXMFyy1m9DqwvDZVbTqjGgsrBzcfOr22ujoVsZqcAC8AfTKqYoqJp66+vjE-iYAKhAACkRQrjAAzsGh8u60dFqh6o1pWWryWt5o8anlMGIAPEVwoHjY1LUmAHzJcGMTrtNwRNhscAD8dSzYajxFLSxgFVWHANq88bwAuuMgk6sA1sBsEASpJeWVuLU7xTKZz+FyKVxut3uPGwAFdJJIvHBocAAG74LToSTAeAmQZiBYkSREaq1cwhPYHAZEIbDOZXfa0O7HZinX4wdp42lqWZXVlVJmxcnMJzYaowKAwkgwaAACgZEVx1JGdN48ruABo4Hy3FSaSrtTA7gBKXahSwSaoAOnlC3lzLNQ0wVoNCwN9p0SAxaCx8HWbAAagsAAxrWp+r0AegjcGoTvFmAARjCYJg0egZX7A8Tdcq1AxOgpZibqtSnQRMMAZmpOSY6PMC8Aa11ZloMxss1XaE2mvWc1zC8XS9Vy5W+7X0iZ5C3Bf0ANJvD4AVWwmGctPmKQoDyeYY22zgr3eny3yLRUC9-WeC3nR+Xq+wwxl44b8iNs0FtEJZzgBBh2Cl95wAA5tiwylNuKydo2Sr9sAsyarOEFTLUACiICEjCtDDDeS4rmupT0modzwXwNyzLMMqnEMPClJqh48LORo0Uhqx+vuiHLMhB4Lp8BGqkR9w7Hxaq3Bcs73JxqwmPuio0sJAlibc8ynvgSJwNgqKqSp56Cj6cCVNUACMCwgTAsHdvQJG8OIUjULwFGZqGY55lZNnSLwRpaHpBkAEwmaBsm5oWmrWRI7kOe2TmBbBdYhW5dmeZi2L6ZWADM-lmdFz4ZK5YV2RFAZRdWMHZZOuW2R5XnJQZAAsGVtoV2ZZXmE7NnFeWVegQA

Currently I have 20 type errors, hopefully I can get it down to 0.