@nicolo-ribaudo 2024-04-19
Oh, it was `11s` because it stopped checking after the first project due to the type errors :sweat_smile: