@nicolo-ribaudo 2024-04-22
Type checking now runs with `skipLibCheck: true`, except for in the root project which:
- uses `skipLibCheck: false`
- checks that all the generated `.d.ts` files are valid

This also caught some errors in the `.d.ts` files ([658b19c](https://github.com/babel/babel/pull/16430/commits/658b19c7f19c95bd6618211262091861ce1a78ff) and [e2ebc5e](https://github.com/babel/babel/pull/16430/commits/e2ebc5eaedd2d6a20dd3ac80759f646ea1f179c9)), that TS was not reporting because it was assuming that the `.d.ts` files were valid.

It takes ~30 seconds to run, which is acceptable.