@nicolo-ribaudo 2024-04-22
`yarn eslint packages --format codeframe` is crashing with OOM, and even when I set `--max-old-space-size=8196` it takes 1m22s. This is _after_ running `make tscheck`, so that all the `.d.ts` and `.tsbuildinfo` files are pre-generated.

cc @bradzacher @JoshuaKGoldberg I assume that this is just because we have 140 projects, but I'm pinging you just in case you want to take a look :) Linting is already slow for us, so this is not a blocker.