@jakebailey 2024-04-22
I was going to suggest `EXPERIMENTAL_projectService`, but it looks like you've already set that... That should have helped with the OOM (and maybe it does but is not enough). 🙁 

Given ESLint doesn't tell plugins when it's done with files, perhaps this is a case of every single file being left open. I tried to take a pprof profile of the run to see where all the memory is going, but that segfaults...