@nicolo-ribaudo 2024-04-22
`NODE_OPTIONS="--max-old-space-size=12294" yarn eslint scripts benchmark packages codemods eslint '*.{js,cjs,mjs,ts}' --format codeframe` is enough, but it's surprising that when using project references the required memory goes from the default 1.5GB to 12GB :grimacing: 

> Given ESLint doesn't tell plugins when it's done with files, perhaps this is a case of every single file being left open.

This shouldn't be affected by using multiple projects, right?