@nicolo-ribaudo 2024-04-22
This PR is ready for review!

Some info:
- a complete type checking takes about 30s both in this PR and on `main`
- full type checking + linting is now a bit slower (1m20s on `main`, 1m40s with this PR)
  - `make lint` now also does type checking to ensure that all the TS projects are pre-compiled, but it will only perform an incremental compilation
- `make tscheck` now also validates all the generated `.d.ts` files, to make sure that there wasn't a suppressed error that made it way to `.d.ts`. This will be helpful when we publish `.d.ts` files
  - this means that the manually-written `.d.ts` file will be checked as part of this step, and we can in other cases run with `skipLibCheck: true`
- `make tscheck` is not exactly the same as `yarn tsc -b .`, because it also copies manually-written `.d.ts` files into the `dts` folder. It needs to be done interleaved with type checking, so that they all all there when checking the `.d.ts` folder. Unfortunately TSC doesn't support this (see https://github.com/microsoft/TypeScript/issues/39231)
- When editing a file locally, incremental type checking now works. For example, if I add `let x: string = 2` to `packages/babel-helper-module-transforms/src/index.ts`, `yarn tsc -b .` only takes 2.1s :rocket: 
- As a rule of thumb, use `yarn tsc -b .` when doing an incremental check, and `make tscheck` when you are not sure (or when you need some manually-written `.d.ts` to be re-copied)