@liuxingbaoyu 2024-04-22
I'm not sure if having `Make types faster` even without using Project References will be enough, but now that we've done that it's good to merge it. :)
lint is a bit strangely slow in this case, but it's acceptable for me.
This may affect the `eslint-vscode` extension, if you are using it you might want to take a look,
Also this might unlock `strictNullChecks`, which I remember would OOM in the past, which is a bonus!