@nicolo-ribaudo 2024-04-22
`eslint-vscode` is not affected by the OOM problem because it only runs on the currently open files :)