@nicolo-ribaudo 2024-04-22
> Also this might unlock strictNullChecks, which I remember would OOM in the past, which is a bonus!

We can now also migrate to `strictNullChecks` project-by-project, rather than having to do everything at the same time!